pyproject.toml
[project]
name = "ai-ecosystem-tracker"
version = "0.1.0"
description = "Automated tracker for AI crawler docs, content ecosystem, and agent infrastructure."
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "feedparser>=6.0",
    "readability-lxml>=0.8.1",
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    "anthropic>=0.40",
    "pyyaml>=6.0",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
    "jinja2>=3.1",
    "google-genai>=1.73.1",
]

[dependency-groups]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "ruff>=0.5",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
